safari-600.7-branch/LayoutTests/fast/dom/HTMLDataGridElement/DataGridColumns-dom-attributes-expected.txt

PASS: Initial value of datagrid.columns is defined.
PASS: Initial length of datagrid.columns is 2.
PASS: The column's ID is from as expected.
PASS: The column's label is From as expected.
PASS: The column's type is text as expected.
PASS: The column's primary property is true as expected.
PASS: The column's sortable property is 2 as expected.
PASS: The first column is the primary column.
